Moved default library to LibraryTab
