+++ src/main/java/org/jabref/gui/preferences/appearance/AppearanceTab.fxml
@@ -11,7 +11,6 @@
 <?import javafx.scene.layout.ColumnConstraints?>
 <?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.layout.VBox?>
 <?import org.jabref.gui.icon.JabRefIconView?>
 <fx:root spacing="10.0" type="VBox"
@@ -23,11 +22,6 @@
             <ColumnConstraints hgrow="SOMETIMES" percentWidth="30.0"/>
             <ColumnConstraints hgrow="SOMETIMES"/>
         </columnConstraints>
-        <rowConstraints>
-            <RowConstraints minHeight="10.0" vgrow="SOMETIMES"/>
-            <RowConstraints minHeight="10.0" vgrow="SOMETIMES"/>
-            <RowConstraints minHeight="10.0" vgrow="SOMETIMES"/>
-        </rowConstraints>
 
         <Label text="%Language"
                GridPane.columnIndex="0" GridPane.columnSpan="2"/>
