+++ src/main/java/org/jabref/gui/preferences/entry/EntryTab.fxml
@@ -3,32 +3,20 @@
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.CheckBox?>
-<?import javafx.scene.control.ComboBox?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.control.Tooltip?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
 <fx:root spacing="10.0" type="VBox"
          xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
          fx:controller="org.jabref.gui.preferences.entry.EntryTab">
     <Label styleClass="titleHeader" text="%Entry"/>
-    <GridPane hgap="10.0" vgap="4.0">
-    <columnConstraints>
-        <ColumnConstraints hgrow="SOMETIMES" percentWidth="30.0"/>
-        <ColumnConstraints hgrow="SOMETIMES"/>
-    </columnConstraints>
-        <Label text="%Default library mode" GridPane.rowIndex="0"/>
-        <ComboBox fx:id="biblatexMode" prefWidth="200.0" GridPane.rowIndex="0" GridPane.columnIndex="1"/>
 
-        <Label styleClass="sectionHeader" text="%Field" GridPane.rowIndex="1" GridPane.columnSpan="2"/>
+    <Label styleClass="sectionHeader" text="%Field"/>
 
-        <Label text="%Keyword separator" GridPane.rowIndex="2"/>
-        <TextField fx:id="keywordSeparator" minWidth="30.0" maxWidth="30.0" alignment="CENTER"
-            GridPane.rowIndex="2" GridPane.columnIndex="1"/>
-    </GridPane>
+    <Label text="%Keyword separator"/>
+    <TextField fx:id="keywordSeparator" minWidth="30.0" maxWidth="30.0" alignment="CENTER"/>
 
     <CheckBox fx:id="resolveStrings" text="%Resolve BibTeX strings"/>
     <HBox alignment="CENTER_LEFT" spacing="10.0">
