+++ src/main/java/org/jabref/gui/preferences/entry/EntryTabViewModel.java
@@ -1,28 +1,19 @@
 package org.jabref.gui.preferences.entry;
 
 import javafx.beans.property.BooleanProperty;
-import javafx.beans.property.ListProperty;
-import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleBooleanProperty;
-import javafx.beans.property.SimpleListProperty;
-import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
-import javafx.collections.FXCollections;
 
 import org.jabref.gui.preferences.PreferenceTabViewModel;
 import org.jabref.logic.bibtex.FieldPreferences;
 import org.jabref.logic.preferences.OwnerPreferences;
 import org.jabref.logic.preferences.TimestampPreferences;
-import org.jabref.model.database.BibDatabaseMode;
 import org.jabref.model.entry.field.FieldFactory;
 import org.jabref.preferences.BibEntryPreferences;
-import org.jabref.preferences.GeneralPreferences;
 import org.jabref.preferences.PreferencesService;
 
 public class EntryTabViewModel implements PreferenceTabViewModel {
-    private final ListProperty<BibDatabaseMode> bibliographyModeListProperty = new SimpleListProperty<>();
-    private final ObjectProperty<BibDatabaseMode> selectedBiblatexModeProperty = new SimpleObjectProperty<>();
 
     private final StringProperty keywordSeparatorProperty = new SimpleStringProperty("");
 
@@ -40,21 +31,16 @@ public class EntryTabViewModel implements PreferenceTabViewModel {
     private final BibEntryPreferences bibEntryPreferences;
     private final OwnerPreferences ownerPreferences;
     private final TimestampPreferences timestampPreferences;
-    private final GeneralPreferences generalPreferences;
 
     public EntryTabViewModel(PreferencesService preferencesService) {
         this.bibEntryPreferences = preferencesService.getBibEntryPreferences();
         this.fieldPreferences = preferencesService.getFieldPreferences();
         this.ownerPreferences = preferencesService.getOwnerPreferences();
         this.timestampPreferences = preferencesService.getTimestampPreferences();
-        this.generalPreferences = preferencesService.getGeneralPreferences();
     }
 
     @Override
     public void setValues() {
-        bibliographyModeListProperty.setValue(FXCollections.observableArrayList(BibDatabaseMode.values()));
-        selectedBiblatexModeProperty.setValue(generalPreferences.getDefaultBibDatabaseMode());
-
         keywordSeparatorProperty.setValue(bibEntryPreferences.getKeywordSeparator().toString());
 
         resolveStringsProperty.setValue(fieldPreferences.shouldResolveStrings());
@@ -71,8 +57,6 @@ public class EntryTabViewModel implements PreferenceTabViewModel {
 
     @Override
     public void storeSettings() {
-        generalPreferences.setDefaultBibDatabaseMode(selectedBiblatexModeProperty.getValue());
-
         bibEntryPreferences.keywordSeparatorProperty().setValue(keywordSeparatorProperty.getValue().charAt(0));
 
         fieldPreferences.setResolveStrings(resolveStringsProperty.getValue());
@@ -85,14 +69,6 @@ public class EntryTabViewModel implements PreferenceTabViewModel {
 
         timestampPreferences.setAddCreationDate(addCreationDateProperty.getValue());
         timestampPreferences.setAddModificationDate(addModificationDateProperty.getValue());
-    }
-
-    public ListProperty<BibDatabaseMode> biblatexModeListProperty() {
-        return this.bibliographyModeListProperty;
-    }
-
-    public ObjectProperty<BibDatabaseMode> selectedBiblatexModeProperty() {
-        return this.selectedBiblatexModeProperty;
     }
 
     public StringProperty keywordSeparatorProperty() {
