+++ src/main/java/org/jabref/gui/preferences/library/LibraryTab.fxml
@@ -1,12 +1,21 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import javafx.scene.control.ComboBox?>
 <?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.ColumnConstraints?>
+<?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.VBox?>
 <fx:root spacing="10.0" type="VBox"
          xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
          fx:controller="org.jabref.gui.preferences.library.LibraryTab">
-    <Label styleClass="titleHeader" text="%General"/>
+    <Label styleClass="titleHeader" text="%Libraries"/>
+    <GridPane hgap="10.0" vgap="4.0">
+        <columnConstraints>
+            <ColumnConstraints hgrow="SOMETIMES" percentWidth="30.0"/>
+            <ColumnConstraints hgrow="SOMETIMES"/>
+        </columnConstraints>
 
-
-
+        <Label text="%Default library mode"/>
+        <ComboBox fx:id="biblatexMode" prefWidth="200.0" GridPane.columnIndex="1"/>
+    </GridPane>
 </fx:root>
