+++ src/main/java/org/jabref/gui/preferences/library/LibraryTab.java
@@ -1,14 +1,19 @@
 package org.jabref.gui.preferences.library;
 
+import javafx.fxml.FXML;
+import javafx.scene.control.ComboBox;
+
 import org.jabref.gui.preferences.AbstractPreferenceTabView;
 import org.jabref.gui.preferences.PreferencesTab;
+import org.jabref.gui.util.ViewModelListCellFactory;
 import org.jabref.logic.l10n.Localization;
+import org.jabref.model.database.BibDatabaseMode;
 
 import com.airhacks.afterburner.views.ViewLoader;
 
 public class LibraryTab extends AbstractPreferenceTabView<LibraryTabViewModel> implements PreferencesTab {
 
-
+    @FXML private ComboBox<BibDatabaseMode> biblatexMode;
 
     public LibraryTab() {
         ViewLoader.view(this)
@@ -23,5 +28,11 @@ public class LibraryTab extends AbstractPreferenceTabView<LibraryTabViewModel> i
 
     public void initialize() {
         this.viewModel = new LibraryTabViewModel(dialogService, preferencesService.getGeneralPreferences(), preferencesService.getTelemetryPreferences());
+
+        new ViewModelListCellFactory<BibDatabaseMode>()
+                .withText(BibDatabaseMode::getFormattedName)
+                .install(biblatexMode);
+        biblatexMode.itemsProperty().bind(viewModel.biblatexModeListProperty());
+        biblatexMode.valueProperty().bindBidirectional(viewModel.selectedBiblatexModeProperty());
     }
 }
