+++ src/main/java/org/jabref/gui/preferences/library/LibraryTabViewModel.java
@@ -1,26 +1,44 @@
 package org.jabref.gui.preferences.library;
 
+import javafx.beans.property.ListProperty;
+import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleListProperty;
+import javafx.beans.property.SimpleObjectProperty;
+import javafx.collections.FXCollections;
+
 import org.jabref.gui.DialogService;
 import org.jabref.gui.preferences.PreferenceTabViewModel;
+import org.jabref.model.database.BibDatabaseMode;
 import org.jabref.preferences.GeneralPreferences;
 import org.jabref.preferences.TelemetryPreferences;
 
 public class LibraryTabViewModel implements PreferenceTabViewModel {
 
+    private final ListProperty<BibDatabaseMode> bibliographyModeListProperty = new SimpleListProperty<>();
+    private final ObjectProperty<BibDatabaseMode> selectedBiblatexModeProperty = new SimpleObjectProperty<>();
+
     private final DialogService dialogService;
     private final GeneralPreferences generalPreferences;
 
-
-    @SuppressWarnings("ReturnValueIgnored")
     public LibraryTabViewModel(DialogService dialogService, GeneralPreferences generalPreferences, TelemetryPreferences telemetryPreferences) {
         this.dialogService = dialogService;
         this.generalPreferences = generalPreferences;
     }
 
     public void setValues() {
+        bibliographyModeListProperty.setValue(FXCollections.observableArrayList(BibDatabaseMode.values()));
+        selectedBiblatexModeProperty.setValue(generalPreferences.getDefaultBibDatabaseMode());
     }
 
     public void storeSettings() {
-        
+        generalPreferences.setDefaultBibDatabaseMode(selectedBiblatexModeProperty.getValue());
+    }
+
+    public ListProperty<BibDatabaseMode> biblatexModeListProperty() {
+        return this.bibliographyModeListProperty;
+    }
+
+    public ObjectProperty<BibDatabaseMode> selectedBiblatexModeProperty() {
+        return this.selectedBiblatexModeProperty;
     }
 }
