+++ src/main/resources/l10n/JabRef_en.properties
@@ -1694,6 +1694,7 @@ Check\ the\ event\ log\ to\ see\ all\ notifications=Check the event log to see a
 Host=Host
 Port=Port
 Library=Library
+Libraries=Libraries
 User=User
 Connect=Connect
 Connection\ error=Connection error
